requirements.txt
django
djangorestframework
django-cors-headers
pyjwt
Pillow
gunicorn